+++ printcetnrum_web.client/src/app/models/design-files-models/stamp-container.model.ts
