+++ printcetnrum_web.client/src/app/pages/user-profile-page/user-profile-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, inject, OnInit } from '@angular/core';
-import { User } from '../../models/user.model';
+import { User } from '../../models/user-models/user.model';
 import { AuthService } from '../../services/auth-services/auth.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { SnackBarUtil } from '../../shared/snackbar-util';
