+++ printcetnrum_web.client/src/app/user-accounts/reset-password-page/reset-password-page.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { LoginPageComponent } from '../../pages/login-page/login-page.component';
 import { ActivatedRoute, Router } from '@angular/router';
-import { ResetPassword } from '../../models/reset-password.model';
+import { ResetPassword } from '../../models/user-models/reset-password.model';
 import { ResetPasswordService } from '../../services/auth-services/reset-password.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { SnackBarUtil } from '../../shared/snackbar-util';
